Edit post: Move CopyContentMenuItem to hooks subfolder
